// apps/mikan-sync/src/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * A signed eight-byte integer. The upper big integer values are greater than the
   * max value for a JavaScript number. Therefore all big integers will be output as
   * strings and not numbers.
   */
  BigInt: any;
  /** A location in a connection that can be used for resuming pagination. */
  Cursor: any;
  /**
   * A point in time as described by the [ISO
   * 8601](https://en.wikipedia.org/wiki/ISO_8601) standard. May or may not include a timezone.
   */
  Datetime: string;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { [key: string]: any };
};
// ...
export type CreateRssItemPayload = {
  __typename?: 'CreateRssItemPayload';
  /**
   * The exact same `clientMutationId` that was provided in the mutation input,
   * unchanged and unused. May be used by a client to track mutations.
   */
  clientMutationId?: Maybe<Scalars['String']>;
  /** Our root query field type. Allows us to run any query from our mutation payload. */
  query?: Maybe<Query>;
  /** The `RssItem` that was created by this mutation. */
  rssItem?: Maybe<RssItem>;
  /** An edge for our `RssItem`. May be used by Relay 1. */
  rssItemEdge?: Maybe<RssItemsEdge>;
};
// ...
export type Mutation = {
  __typename?: 'Mutation';
  /** Creates a single `AnimeMetadatum`. */
  createAnimeMetadatum?: Maybe<CreateAnimeMetadatumPayload>;
  /** Creates a single `DownloadConfig`. */
  createDownloadConfig?: Maybe<CreateDownloadConfigPayload>;
  /** Creates a single `RssItem`. */
  createRssItem?: Maybe<CreateRssItemPayload>;
  /** Creates a single `SonarrSery`. */
  createSonarrSery?: Maybe<CreateSonarrSeryPayload>;
  /** Deletes a single `AnimeMetadatum` using its globally unique id. */
  deleteAnimeMetadatum?: Maybe<DeleteAnimeMetadatumPayload>;
  /** Deletes a single `AnimeMetadatum` using a unique key. */
  deleteAnimeMetadatumById?: Maybe<DeleteAnimeMetadatumPayload>;
  /** Deletes a single `DownloadConfig` using its globally unique id. */
  deleteDownloadConfig?: Maybe<DeleteDownloadConfigPayload>;
  /** Deletes a single `DownloadConfig` using a unique key. */
  deleteDownloadConfigById?: Maybe<DeleteDownloadConfigPayload>;
  /** Deletes a single `RssItem` using its globally unique id. */
  deleteRssItem?: Maybe<DeleteRssItemPayload>;
  /** Deletes a single `RssItem` using a unique key. */
  deleteRssItemByHash?: Maybe<DeleteRssItemPayload>;
  /** Deletes a single `RssItem` using a unique key. */
  deleteRssItemById?: Maybe<DeleteRssItemPayload>;
  /** Deletes a single `SonarrSery` using its globally unique id. */
  deleteSonarrSery?: Maybe<DeleteSonarrSeryPayload>;
  /** Deletes a single `SonarrSery` using a unique key. */
  deleteSonarrSeryById?: Maybe<DeleteSonarrSeryPayload>;
  /** Deletes a single `SonarrSery` using a unique key. */
  deleteSonarrSeryBySonarrId?: Maybe<DeleteSonarrSeryPayload>;
  /** Deletes a single `SonarrSery` using a unique key. */
  deleteSonarrSeryBySonarrSlug?: Maybe<DeleteSonarrSeryPayload>;
  /** Updates a single `AnimeMetadatum` using its globally unique id and a patch. */
  updateAnimeMetadatum?: Maybe<UpdateAnimeMetadatumPayload>;
  /** Updates a single `AnimeMetadatum` using a unique key and a patch. */
  updateAnimeMetadatumById?: Maybe<UpdateAnimeMetadatumPayload>;
  /** Updates a single `DownloadConfig` using its globally unique id and a patch. */
  updateDownloadConfig?: Maybe<UpdateDownloadConfigPayload>;
  /** Updates a single `DownloadConfig` using a unique key and a patch. */
  updateDownloadConfigById?: Maybe<UpdateDownloadConfigPayload>;
  /** Updates a single `RssItem` using its globally unique id and a patch. */
  updateRssItem?: Maybe<UpdateRssItemPayload>;
  /** Updates a single `RssItem` using a unique key and a patch. */
  updateRssItemByHash?: Maybe<UpdateRssItemPayload>;
  /** Updates a single `RssItem` using a unique key and a patch. */
  updateRssItemById?: Maybe<UpdateRssItemPayload>;
  /** Updates a single `SonarrSery` using its globally unique id and a patch. */
  updateSonarrSery?: Maybe<UpdateSonarrSeryPayload>;
  /** Updates a single `SonarrSery` using a unique key and a patch. */
  updateSonarrSeryById?: Maybe<UpdateSonarrSeryPayload>;
  /** Updates a single `SonarrSery` using a unique key and a patch. */
  updateSonarrSeryBySonarrId?: Maybe<UpdateSonarrSeryPayload>;
  /** Updates a single `SonarrSery` using a unique key and a patch. */
  updateSonarrSeryBySonarrSlug?: Maybe<UpdateSonarrSeryPayload>;
};
// ...
export type Query = Node & {
  __typename?: 'Query';
  /** Reads and enables pagination through a set of `AnimeMetadatum`. */
  allAnimeMetadata?: Maybe<AnimeMetadataConnection>;
  /** Reads and enables pagination through a set of `DownloadConfig`. */
  allDownloadConfigs?: Maybe<DownloadConfigsConnection>;
  /** Reads and enables pagination through a set of `RssItem`. */
  allRssItems?: Maybe<RssItemsConnection>;
  /** Reads and enables pagination through a set of `SonarrSery`. */
  allSonarrSeries?: Maybe<SonarrSeriesConnection>;
  /** Reads a single `AnimeMetadatum` using its globally unique `ID`. */
  animeMetadatum?: Maybe<AnimeMetadatum>;
  animeMetadatumById?: Maybe<AnimeMetadatum>;
  /** Reads a single `DownloadConfig` using its globally unique `ID`. */
  downloadConfig?: Maybe<DownloadConfig>;
  downloadConfigById?: Maybe<DownloadConfig>;
  /** Fetches an object given its globally unique `ID`. */
  node?: Maybe<Node>;
  /** The root query type must be a `Node` to work well with Relay 1 mutations. This just resolves to `query`. */
  nodeId: Scalars['ID'];
  /**
   * Exposes the root query type nested one level down. This is helpful for Relay 1
   * which can only query top level fields if they are in a particular form.
   */
  query: Query;
  /** Reads a single `RssItem` using its globally unique `ID`. */
  rssItem?: Maybe<RssItem>;
  rssItemByHash?: Maybe<RssItem>;
  rssItemById?: Maybe<RssItem>;
  /** Reads a single `SonarrSery` using its globally unique `ID`. */
  sonarrSery?: Maybe<SonarrSery>;
  sonarrSeryById?: Maybe<SonarrSery>;
  sonarrSeryBySonarrId?: Maybe<SonarrSery>;
  sonarrSeryBySonarrSlug?: Maybe<SonarrSery>;
};
// ...
export type RssItem = Node & {
  __typename?: 'RssItem';
  hash: Scalars['String'];
  id: Scalars['Int'];
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID'];
  publishDate: Scalars['Datetime'];
  size: Scalars['BigInt'];
  title: Scalars['String'];
  torrentLink: Scalars['String'];
};
// ...
export type RssItemInput = {
  hash: Scalars['String'];
  id?: InputMaybe<Scalars['Int']>;
  publishDate: Scalars['Datetime'];
  size: Scalars['BigInt'];
  title: Scalars['String'];
  torrentLink: Scalars['String'];
};
// ...
export type RssItemsConnection = {
  __typename?: 'RssItemsConnection';
  /** A list of edges which contains the `RssItem` and cursor to aid in pagination. */
  edges: Array<RssItemsEdge>;
  /** A list of `RssItem` objects. */
  nodes: Array<Maybe<RssItem>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `RssItem` you could get from the connection. */
  totalCount: Scalars['Int'];
};

/** A `RssItem` edge in the connection. */
export type RssItemsEdge = {
  __typename?: 'RssItemsEdge';
  /** A cursor for use in pagination. */
  cursor?: Maybe<Scalars['Cursor']>;
  /** The `RssItem` at the end of the edge. */
  node?: Maybe<RssItem>;
};
// ...
export type GetExistingRssItemsWithHashQueryVariables = Exact<{
  in: Array<Scalars['String']> | Scalars['String'];
}>;


export type GetExistingRssItemsWithHashQuery = { __typename?: 'Query', allRssItems?: { __typename?: 'RssItemsConnection', edges: Array<{ __typename?: 'RssItemsEdge', node?: { __typename?: 'RssItem', id: number, publishDate: string, hash: string, size: any, title: string, torrentLink: string } | null | undefined }> } | null | undefined };

export type CreateRssItemMutationVariables = Exact<{
  rssItem: RssItemInput;
}>;


export type CreateRssItemMutation = { __typename?: 'Mutation', createRssItem?: { __typename: 'CreateRssItemPayload' } | null | undefined };


export const GetExistingRssItemsWithHashDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"GetExistingRSSItemsWithHash"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"in"}},"type":{"kind":"NonNullType","type":{"kind":"ListType","type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"allRssItems"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"filter"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"hash"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"in"},"value":{"kind":"Variable","name":{"kind":"Name","value":"in"}}}]}}]}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"edges"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"node"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"publishDate"}},{"kind":"Field","name":{"kind":"Name","value":"hash"}},{"kind":"Field","name":{"kind":"Name","value":"size"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"torrentLink"}}]}}]}}]}}]}}]} as unknown as DocumentNode<GetExistingRssItemsWithHashQuery, GetExistingRssItemsWithHashQueryVariables>;
export const CreateRssItemDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"mutation","name":{"kind":"Name","value":"CreateRSSItem"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"rssItem"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"RssItemInput"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"createRssItem"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"rssItem"},"value":{"kind":"Variable","name":{"kind":"Name","value":"rssItem"}}}]}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}}]}}]}}]} as unknown as DocumentNode<CreateRssItemMutation, CreateRssItemMutationVariables>;